+++ src/main/kotlin/common/ui/helpers/swingModelHelpers.kt
@@ -1,7 +1,6 @@
 package common.ui.helpers
 
 import com.intellij.ui.DocumentAdapter
-import javax.swing.JComboBox
 import javax.swing.JList
 import javax.swing.event.DocumentEvent
 import javax.swing.text.Document
@@ -13,17 +12,6 @@ fun <T> JList<T>.onSelectedValue(action: JList<T>.(T?) -> Unit) {
         if (prevValue != selectedValue)
             action(selectedValue)
         prevValue = selectedValue
-    }
-}
-
-@Suppress("UNCHECKED_CAST")
-fun <T> JComboBox<T>.onSelectedItem(action: JComboBox<T>.(T?) -> Unit) {
-    var prevValue: T? = null
-    action(selectedItem as T?)
-    addActionListener {
-        if (prevValue != selectedItem)
-            action(selectedItem as T?)
-        prevValue = selectedItem as T?
     }
 }