+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 ?> All notable changes to this project will be documented in this page.
 
 ## [Unreleased]
+
+### Added
+
+- Build Configuration Macros: ```@in```, ```@dir``` and ```$dir```
+
+### Changed
+
+- Build Configuration commandTemplate split into Build Command and Execute Command
+
 ## v0.5.0
 
 ### Added